add: slider component
