+++ lib/components/Slider/styles.css
